autosales-backend-lite/requirements.txt
fastapi
uvicorn
psycopg2-binary
requests
pydantic
python-dotenv
python-multipart
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose
pydantic[email]